+++ src/main/kotlin/com/transportes/dto/payment/PaymentInfoDTO.kt
@@ -1,9 +1,11 @@
 package com.transportes.dto.payment
 
+import java.math.BigDecimal
+
 data class PaymentInfoDTO(
     val publisher: String,
     val transport: String,
     val origin: String,
     val destination: String,
-    val mount: Double
+    val mount: BigDecimal
 )
